moved testable functions to a seperate file
