+++ connect4.test.js
@@ -1,56 +1,61 @@
 // check cols
+const { checkCols, getBoard } = require('./functions');
 const each = require("jest-each").default;
 
-function checkCols(row, column) {
-    if (row < connectN - 1) { // leave if too close to bottom of board
-        return false;
-    }
-    for (let i = 1; i < connectN; i++) {
-        if (board[row][column] === board[row - i][column]) {
-            if (i === connectN - 1) {
-                console.log('column winner')
-                return true;
-            }
-            continue;
-        } else {
-            break;
-        }
-    }
-    return false;
-}
-
-function getBoard() {
-    let board = new Array(rows);
-    for (let i = 0; i < rows; i++) {
-        board[i] = new Array(cols);
-        board[i].fill(null)
-    }
-    return board;
-}
-
-describe("multi_test", () => {
+describe("token in empty board", () => {
     rows = 6;
     cols = 7;
     connectN = 4;
 
-    let t1_board = getBoard();
-    t1_board[0][3] = 'y'
-    let t1_counter = [0, 3]
-    let t1_expected = false;
 
-
+    let t1_counter = 'y'
+    let t2_counter = 'r'
+    let expected = false;
 
     each([
-        ["Place in empty board", t1_board, t1_counter, t1_expected],
+        ["Place yellow in empty board", t1_counter, expected],
+        ["Place red in empty board", t2_counter, expected],
 
 
-    ]).it("'%s'", (text, board, last_placed_counter, expected_output) => {
+    ]).it("'%s'", (text, counter, expected_output) => {
 
-        expect(checkCols(t1_counter[0], t1_counter[1])).toStrictEqual(false);
+        let board = getBoard();
+        let placedCounter = [0, 0]
+        board[0][0] = counter;
+        expect(checkCols(placedCounter[0], placedCounter[1])).toStrictEqual(expected_output);
+
 
     });
 });
 
+describe("column in empty board", () => {
+    rows = 6;
+    cols = 7;
+    connectN = 4;
+
+
+    let expected = true;
+    counter1 = 'y'
+    counter2 = 'r'
+
+    each([
+        ["yellow column in empty board", counter1, expected],
+        ["red column in empty board", counter2, expected],
+
+    ]).it("'%s'", (text, counter, expected_output) => {
+
+        let board = getBoard();
+        board[0][0] = counter
+        board[1][0] = counter
+        board[2][0] = counter
+        board[3][0] = counter
+
+        let placedCounter = [3][0]
+
+        expect(checkCols(placedCounter[0], placedCounter[1])).toStrictEqual(expected_output);
+
+    });
+});
 
 // check rows
