+++ index.html
@@ -24,6 +24,7 @@
 
     </div>
 
+    <script src="functions.js" type="text/javascript"></script>0
     <script src="connect4.js" type="text/javascript"></script>
